backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
networkx
google-generativeai>=0.3.0
python-dotenv
pytest
httpx
PyPDF2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-socketio
aiofiles


PyJWT==2.8.0


# Additional utilities
python-json-logger>=2.0.0
numpy>=1.24.0

# Security & Encryption
cryptography>=41.0.0

# Content Providers
requests>=2.31.0
youtube-transcript-api>=0.6.0
